Add TKG Service 3.3.1 and 3.3.2
Add links to the package.yaml and open_source_license.txt for each.

Refactor the way we refer to the interop matrix to avoid duplication; it
seems most useful to have a single link prior to the version list as a
reader may use compatibility information to understand which version(s)
meet their needs.

Additionally, include instructions for finding the End of Service date
for Kubernetes releases. This seems like useful context for a reader to
consider in addition to the compatibility information.